+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - CircleCI support
 - Enforce updates to `CHANGELOG.md` for PRs
+- Support bitbucket repos
 
 ## [2.1.0][] - 2017-05-25